+++ bin/tab_completion.sh
@@ -50,6 +50,8 @@ _mde() {
             
               --dump-selected-block) COMPREPLY="0"; return 0 ;;
             
+              -w) COMPREPLY="0"; return 0 ;;
+            
               --filename) COMPREPLY="."; return 0 ;;
             
               -f) COMPREPLY="."; return 0 ;;
@@ -67,6 +69,8 @@ _mde() {
               --list-count) COMPREPLY="32"; return 0 ;;
             
               --load-code) COMPREPLY="''"; return 0 ;;
+            
+              -l) COMPREPLY="''"; return 0 ;;
             
               --open) COMPREPLY="''"; return 0 ;;
             
@@ -94,7 +98,7 @@ _mde() {
   # present matching option names
   #
   if [[ ${cur} == -* ]] ; then
-    opts=("--block-name" "--config" "--debug" "--dump-delegate-object" "--dump-blocks-in-file" "--dump-inherited-block_names" "--dump-inherited-dependencies" "--dump-inherited-lines" "--dump-menu-blocks" "--dump-selected-block" "--exit" "--filename" "--find" "--find-path" "--help" "--how" "--list-blocks" "--list-count" "--list-default-env" "--list-default-yaml" "--list-docs" "--load-code" "--open" "--output-script" "--output-stdout" "--path" "--pwd" "--run-last-script" "--select-recent-output" "--select-recent-script" "--tab-completions" "--user-must-approve" "--version" "--display-level")
+    opts=("--block-name" "--config" "--debug" "--dump-delegate-object" "--dump-blocks-in-file" "--dump-inherited-block_names" "--dump-inherited-dependencies" "--dump-inherited-lines" "--dump-menu-blocks" "--dump-selected-block" "--exit" "--filename" "--find" "--find-path" "--help" "--how" "--list-blocks" "--list-count" "--list-default-env" "--list-default-yaml" "--list-docs" "--list-recent-output" "--list-recent-scripts" "--load-code" "--open" "--output-script" "--output-stdout" "--path" "--pwd" "--run-last-script" "--tab-completions" "--user-must-approve" "--version" "--display-level")
     COMPREPLY=( $(compgen -W "$(printf "'%s' " "${opts[@]}")" -- "${cur}") )
 
     return 0
@@ -131,6 +135,8 @@ _mde() {
         
           --dump-selected-block) COMPREPLY=".BOOL."; return 0 ;;
         
+          -w) COMPREPLY=".BOOL."; return 0 ;;
+        
           --filename) COMPREPLY=".RELATIVE_PATH."; return 0 ;;
         
           -f) COMPREPLY=".RELATIVE_PATH."; return 0 ;;
@@ -148,6 +154,8 @@ _mde() {
           --list-count) COMPREPLY=".INT.1-."; return 0 ;;
         
           --load-code) COMPREPLY=".PATH."; return 0 ;;
+        
+          -l) COMPREPLY=".PATH."; return 0 ;;
         
           --open) COMPREPLY=".OPEN."; return 0 ;;